Issue #75 Convert travel_x and travel_y to INT8.
This fixes compile issues, which fails for `signed int`.
It appears the user _can_ travel in all directions, but travel is very
'buggy', particularly when travelling diagonally
-e 
Original/build commit ID: 2708f42a10ad2eb73b18443678be5f48f981d104

+++ joy.h
@@ -9,8 +9,8 @@
 
 typedef struct {
     // Determine which way user needs to travel
-    signed int travel_x;
-    signed int travel_y;
+    INT8 travel_x;
+    INT8 travel_y;
 
     unsigned short a_pressed;
     unsigned short b_pressed;
